backend/requirements.txt
fastapi
uvicorn
aiosqlite
pydantic
python-jose[cryptography]
bcrypt
python-multipart
